+++ risk/forms.py
@@ -18,12 +18,9 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from django import forms
 from django.forms import ModelForm
 
-from risk.Objectives import Playbook
 from risk.Scenarios import Scenario
-from risk.Workflows import Workflow
 
 
 class ScenarioForm(ModelForm):
@@ -31,22 +28,3 @@ class ScenarioForm(ModelForm):
         model = Scenario
         exclude = None
         fields = '__all__'
-
-
-class WorkflowInteractiveForm(ModelForm):
-    class Meta:
-        model = Workflow
-        fields = ['portfolio', 'results_list']
-        widgets = {'results_list': forms.TextInput(attrs={'style': 'width:200px; height:50px;'})}
-
-
-class WorkflowBatchForm(ModelForm):
-    class Meta:
-        model = Workflow
-        fields = ['name']
-
-
-class PlaybookBatchForm(ModelForm):
-    class Meta:
-        model = Playbook
-        fields = ['name']