+++ risk/static/risk/js/plot.js
@@ -1,0 +1,4 @@
+    // TIMESERIES GRAPH
+
+
+    timeseries_graph(dates, values);