+++ risk/CHANGELOG.rst
@@ -1,6 +1,11 @@
 ChangeLog
 ===========================
 
+v0.4.0 (XX-06-2022)
+--------------------
+* Cleanup of unused / undeveloped functionality
+* Full implementation of emission scenarios
+
 v0.3.0 (21-04-2022)
 -------------------
 * Earth Day 2022 release